refactored the quantum solvers
