+++ src/linsolve/_version.py
@@ -1,0 +1,4 @@
+# file generated by setuptools_scm
+# don't change, don't track in version control
+__version__ = version = '1.0.1.dev30+gc96e25c.d20230720'
+__version_tuple__ = version_tuple = (1, 0, 1, 'dev30', 'gc96e25c.d20230720')
