+++ src/linsolve/quantum/__init__.py
@@ -1,0 +1,6 @@
+from .vqls_linsolve import (
+    QuantumLinearSolver,
+    QuantumLinProductSolver,
+    QuantumLogProductSolver,
+)
+from .qubo_linsolve import QUBOLinearSolver, QUBOLinProductSolver, QUBOLogProductSolver
